New folder/requirements.txt
torch
torchvision
torchaudio
numpy
scikit-learn
matplotlib
opencv-python
pillow
tqdm
